Create a Custom Job Balancer sample
This sample implements functionality similar to one of the Deadline 10 balancer options.

Instead of a Priority/FIFO queue, this implements weighted balancing for jobs in the farm.

There are some special edge cases here:
- If there are jobs that are set to 100, the Balancer goes into a "High Priority" mode and the entire Fleet gets balanced only between jobs at 100 priority. Other jobs get paused.
- Jobs at 0 priority do not get any workers UNLESS there are only jobs with priority 0 on the farm. These jobs are treated as "paused"

The actual values for how many workers to assign to each job is done on a "squared weight" policy. That means that a job at priority 20 would get twice as many workers as a job at priority 10.

Signed-off-by: Alex Hughes <ahughesalex@gmail.com>